Begin refactoring data decoding module
